+++ source/gameos/include/networklobby.hpp
@@ -12,7 +12,7 @@
 #include "networking.hpp"
 
 
-typedef void(__stdcall *LPSTRINGCALLBACK)(PCSTR string, DWORD value);
+typedef void(__stdcall *LPSTRINGCALLBACK)(PCSTR string, ULONG value);
 
 
 
@@ -37,7 +37,7 @@ GOSERRORCODE __stdcall gos_EnumerateSessions(LPSTRINGCALLBACK callback);
 // Return value: If successful, return value is GOS_OK.  If no session matching <session_name>
 //					is found, returns GOS_ERR_SESSIONNOTFOUND.
 //
-GOSERRORCODE __stdcall gos_EnumeratePlayers(DWORD session_id, LPSTRINGCALLBACK callback);
+GOSERRORCODE __stdcall gos_EnumeratePlayers(ULONG session_id, LPSTRINGCALLBACK callback);
 
 
 // Functionality: Creates a new game using the given game name, player name and maximum players.
@@ -53,7 +53,7 @@ GOSERRORCODE __stdcall gos_CreateGame(PSTR game_name, PSTR player_name, int max_
 // Return value: If successful, return value is GOS_OK.  If no session matching <session_id>
 //					is found, returns GOS_ERR_SESSIONNOTFOUND.
 //
-GOSERRORCODE __stdcall gos_JoinGame(DWORD session_id, PSTR player_name);
+GOSERRORCODE __stdcall gos_JoinGame(ULONG session_id, PSTR player_name);
 
 
 // Functionality: Calls the callback once for each modem installed on this
